add config

+++ src/Console/Uninstaller.php
@@ -8,11 +8,17 @@ use Illuminate\Support\Facades\DB;
 
 abstract class Uninstaller extends AbstractCommand
 {
+    protected $dropSchema = true;
+
     public function handle(): void
     {
-        $this->call('migrate:reset', ['--path' => migrations_path(static::class)]);
+        $this->call('migrate:reset', [
+            '--path' => migrations_path(static::class),
+        ]);
 
-        DB::statement("DROP SCHEMA IF EXISTS {$this->schemaName()}");
+        if ($this->dropSchema === true) {
+            DB::statement("DROP SCHEMA IF EXISTS {$this->schemaName()}");
+        }
     }
 
     private function schemaName(): string
